Fixed minion spawn and movement
- removed level1map in favor of levelOneMap
- added missing character scaling argument in Entity
