+++ braindefence/arcade/app.py
@@ -25,8 +25,6 @@ class BrainDefence(arcade.Window):
         self.current_map = LevelOneMap()
         self.current_map.render_map()
 
-
-
     def on_draw(self):
         """Render the screen."""
 
@@ -61,9 +59,6 @@ class BrainDefence(arcade.Window):
 
         # print(object_list)
         # mats = arcade.get_sprites_at_point((x, y), sprite_lists)
-
-    def on_update(self, delta_time: float):
-        self.current_map.update(delta_time)
 
     def on_update(self, delta_time: float):
         self.current_map.update(delta_time)
