+++ braindefence/arcade/entities/entity.py
@@ -3,15 +3,15 @@ from braindefence.arcade.constants import *
 
 
 class Entity(arcade.Sprite):
-    def __init__(self, imagefilepath):
-        super().__init__(imagefilepath, CHARACTER_SCALING)
+    def __init__(self, imagefilepath, character_scaling):
+        super().__init__(imagefilepath, character_scaling)
 
         # Default to facing right
         self.facing_direction = RIGHT_FACING
 
         # Used for image sequences
         self.cur_texture = 0
-        self.scale = CHARACTER_SCALING
+        self.scale = character_scaling
         self.character_face_direction = RIGHT_FACING
         self.center_x = World.Width // 2
         self.center_y = World.Height // 2
